+++ packages/model-typings/src/models/ILivechatRoomsModel.ts
@@ -260,4 +260,10 @@ export interface ILivechatRoomsModel extends IBaseModel<IOmnichannelRoom> {
 	getVisitorActiveForPeriodUpdateQuery(period: string, updater?: Updater<IOmnichannelRoom>): Updater<IOmnichannelRoom>;
 	getMACStatisticsForPeriod(period: string): Promise<MACStats[]>;
 	getMACStatisticsBetweenDates(start: Date, end: Date): Promise<MACStats[]>;
+	findPaginatedRoomsByVisitorsIdsAndSource(params: {
+		visitorsIds: string[];
+		source?: string;
+		options?: FindOptions;
+	}): FindPaginated<FindCursor<IOmnichannelRoom>>;
+	countLivechatRoomsWithDepartment(): Promise<number>;
 }